requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
keyboard==0.13.5
openai==1.84.0
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
pynput==1.8.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5-stubs==5.15.6.0
PyQt5_sip==12.17.0
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
